requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels-redis==4.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect2==0.7.0
daphne==4.1.2
Django==5.1.4
django-appconf==1.0.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.15
django-cleanup==9.0.0
django-clearcache==1.2.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-db-connection-pool==1.2.5
django-filter==24.3
django-js-asset==3.0.1
django-redis==5.4.0
django-rest-framework-datatables-filter==0.0.2
django-rest-multiple-models==2.1.3
django-select2==8.2.3
django-simple-history==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-datatables==0.7.0
djangorestframework-datatables-editor==0.3.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
packaging==24.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.3
tinycss2==1.4.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.2
pyodbc==5.2.0
websockets==15.0
websocket-client==1.8.0
pandas==2.2.3
twilio==9.4.6
sendgrid==6.11.0
openpyxl==3.1.5